requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohttp==3.7.4.post0      # via gcsfs
appdirs==1.4.4            # via virtualenv
asciitree==0.3.3          # via zarr
async-timeout==3.0.1      # via aiohttp
attrs==21.2.0             # via aiohttp, pytest
cachetools==4.2.2         # via google-auth
certifi==2020.12.5        # via requests
cfgv==3.2.0               # via pre-commit
chardet==4.0.0            # via aiohttp, requests
click==8.0.0              # via pip-tools
cloudpickle==1.6.0        # via dask
dask==2021.4.1            # via -r requirements.in
decorator==5.0.7          # via gcsfs
distlib==0.3.1            # via virtualenv
fasteners==0.16           # via zarr
filelock==3.0.12          # via tox, virtualenv
fsspec==2021.4.0          # via -r requirements.in, dask, gcsfs
gcsfs==2021.4.0           # via -r requirements.in
google-auth-oauthlib==0.4.4  # via gcsfs
google-auth==1.30.0       # via gcsfs, google-auth-oauthlib
identify==2.2.4           # via pre-commit
idna==2.10                # via requests, yarl
importlib-metadata==4.0.1  # via pep517, pluggy, pre-commit, pytest, tox, virtualenv
iniconfig==1.1.1          # via pytest
locket==0.2.1             # via partd
multidict==5.1.0          # via aiohttp, yarl
nodeenv==1.6.0            # via pre-commit
numcodecs==0.7.3          # via zarr
numpy==1.20.3             # via numcodecs, pandas, xarray, zarr
oauthlib==3.1.0           # via requests-oauthlib
packaging==20.9           # via pytest, tox
pandas==1.2.4             # via xarray
partd==1.2.0              # via dask
pep517==0.10.0            # via pip-tools
pip-tools==6.1.0          # via -r requirements.in
pluggy==0.13.1            # via pytest, tox
pre-commit==2.12.1        # via -r requirements.in
py==1.10.0                # via pytest, tox
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pyparsing==2.4.7          # via packaging
pytest==6.2.4             # via -r requirements.in
python-dateutil==2.8.1    # via pandas
python-dotenv==0.17.1     # via -r requirements.in
pytz==2021.1              # via pandas
pyyaml==5.4.1             # via dask, pre-commit
requests-oauthlib==1.3.0  # via google-auth-oauthlib
requests==2.25.1          # via gcsfs, requests-oauthlib
rsa==4.7.2                # via google-auth
six==1.16.0               # via fasteners, google-auth, python-dateutil, tox, virtualenv
toml==0.10.2              # via pep517, pre-commit, pytest, tox
toolz==0.11.1             # via dask, partd
tox==3.23.1               # via -r requirements.in
typing-extensions==3.10.0.0  # via aiohttp, importlib-metadata, yarl
urllib3==1.26.4           # via requests
virtualenv==20.4.6        # via pre-commit, tox
xarray==0.18.0            # via -r requirements.in
yarl==1.6.3               # via aiohttp
zarr==2.8.1               # via -r requirements.in
zipp==3.4.1               # via importlib-metadata, pep517

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
